Add more type hints, move date functions to utils
